requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
numpy==1.26.2
openpyxl==3.1.2
aiofiles==23.2.1
opencv-python==4.8.1.78
onnxruntime>=1.17.0
apscheduler>=3.10.0
httpx>=0.27.0
